+++ src/main/java/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusLuceneSearchService.java
@@ -39,28 +39,15 @@ public class NexusLuceneSearchService implements IVersionReader {
 	private static final Logger LOGGER = Logger.getLogger(NexusLuceneSearchService.class.getName());
 
 	private final String mURL;
-	private final String mGroupId;
-	private final String mArtifactId;
-	private final String mPackaging;
-	private final ValidAndInvalidClassifier mClassifier;
 
 	private String mUserName;
 	private String mUserPassword;
 
 	private WebResource mInstance;
 
-	public NexusLuceneSearchService(String pURL, String pGroupId, String pArtifactId, String pPackaging) {
-		this(pURL, pGroupId, pArtifactId, pPackaging, ValidAndInvalidClassifier.getDefault());
-	}
-
-	public NexusLuceneSearchService(String pURL, String pGroupId, String pArtifactId, String pPackaging,
-			ValidAndInvalidClassifier pAcceptedClassifier) {
+	public NexusLuceneSearchService(String pURL) {
 		super();
 		this.mURL = pURL;
-		this.mGroupId = pGroupId;
-		this.mArtifactId = pArtifactId;
-		this.mPackaging = pPackaging;
-		this.mClassifier = pAcceptedClassifier;
 	}
 
 	void init() {
@@ -82,27 +69,33 @@ public class NexusLuceneSearchService implements IVersionReader {
 		//
 	}
 
+	public List<String> retrieveVersions(String pGroupId, String pArtifactId, String pPackaging)
+			throws VersionReaderException {
+		return retrieveVersions(pGroupId, pArtifactId, pPackaging, ValidAndInvalidClassifier.getDefault());
+	}
+
 	/**
 	 * Search in Nexus for the artifact using the Lucene Service.
 	 * https://repository.sonatype.org/nexus-indexer-lucene-plugin/default/docs/path__lucene_search.html
 	 */
-	public List<String> retrieveVersions() throws VersionReaderException {
+	public List<String> retrieveVersions(String pGroupId, String pArtifactId, String pPackaging,
+			ValidAndInvalidClassifier pClassifier) throws VersionReaderException {
 		if (LOGGER.isLoggable(Level.FINE)) {
-			LOGGER.fine("query nexus with arguments: r:" + mURL + ", g:" + getGroupId() + ", a:" + getArtifactId()
-					+ ", p:" + getPackaging() + ", c: " + getClassifier().toString());
+			LOGGER.fine("query nexus with arguments: r:" + mURL + ", g:" + pGroupId + ", a:" + pArtifactId + ", p:"
+					+ pPackaging + ", c: " + pClassifier.toString());
 		}
 
 		MultivaluedMap<String, String> requestParams = new MultivaluedHashMap<String, String>();
-		if (getGroupId() != "")
-			requestParams.putSingle("g", getGroupId());
-		if (getArtifactId() != "")
-			requestParams.putSingle("a", getArtifactId());
-		if (getPackaging() != "" && !PACKAGING_ALL.equals(getPackaging()))
-			requestParams.putSingle("p", getPackaging());
-		if (getClassifier() != null) {
+		if (pGroupId != "")
+			requestParams.putSingle("g", pGroupId);
+		if (pArtifactId != "")
+			requestParams.putSingle("a", pArtifactId);
+		if (pPackaging != "" && !PACKAGING_ALL.equals(pPackaging))
+			requestParams.putSingle("p", pPackaging);
+		if (pClassifier != null) {
 			// FIXME: There is of course a better way how to do it...
 			final List<String> query = new ArrayList<String>();
-			for (String current : getClassifier().getValid())
+			for (String current : pClassifier.getValid())
 				query.add(current);
 
 			if (!query.isEmpty())
@@ -122,7 +115,7 @@ public class NexusLuceneSearchService implements IVersionReader {
 				LOGGER.info("response from Nexus does not contain any results.");
 			} else {
 				final Map<String, String> repoURLs = retrieveRepositoryURLs(xmlResult.getRepoDetails());
-				retVal = parseResponse(xmlResult, repoURLs);
+				retVal = parseResponse(xmlResult, repoURLs, pPackaging, pClassifier);
 			}
 		} catch (UniformInterfaceException e) {
 			final String msg;
@@ -140,15 +133,14 @@ public class NexusLuceneSearchService implements IVersionReader {
 			}
 			throw new VersionReaderException(msg, e);
 		} catch (Exception e) {
-			throw new VersionReaderException(
-					"failed to retrieve versions from nexus for r:" + getURL() + ", g:" + getGroupId() + ", a:"
-							+ getArtifactId() + ", p:" + getPackaging() + ", c:" + getClassifier() + e.getMessage(),
-					e);
+			throw new VersionReaderException("failed to retrieve versions from nexus for r:" + getURL() + ", g:"
+					+ pGroupId + ", a:" + pArtifactId + ", p:" + pPackaging + ", c:" + pClassifier + e.getMessage(), e);
 		}
 		return new ArrayList<String>(retVal);
 	}
 
-	Set<String> parseResponse(final PatchedSearchNGResponse pXMLResult, final Map<String, String> pRepoURLs) {
+	Set<String> parseResponse(final PatchedSearchNGResponse pXMLResult, final Map<String, String> pRepoURLs,
+			final String pPackaging, final ValidAndInvalidClassifier pClassifier) {
 		// Use a Map instead of a List to filter duplicated entries and also linked to keep the order of XML response
 		final Set<String> retVal = new LinkedHashSet<String>();
 
@@ -177,16 +169,16 @@ public class NexusLuceneSearchService implements IVersionReader {
 					boolean addCurrentyEntryAsFolder = false;
 
 					// if packaging configuration is set but does not match
-					if ("".equals(getPackaging())) {
+					if ("".equals(pPackaging)) {
 						addCurrentyEntryAsFolder = true;
-					} else if (PACKAGING_ALL.equals(getPackaging())) {
+					} else if (PACKAGING_ALL.equals(pPackaging)) {
 						// then always add
-					} else if (!getPackaging().equals(currentLink.getExtension())) {
+					} else if (!pPackaging.equals(currentLink.getExtension())) {
 						addCurrentEntry &= false;
 					}
 
 					// check the classifier.
-					if (!getClassifier().isValid(currentLink.getClassifier())) {
+					if (!pClassifier.isValid(currentLink.getClassifier())) {
 						addCurrentEntry &= false;
 					}
 
@@ -225,22 +217,6 @@ public class NexusLuceneSearchService implements IVersionReader {
 		return mURL;
 	}
 
-	public String getGroupId() {
-		return mGroupId;
-	}
-
-	public String getArtifactId() {
-		return mArtifactId;
-	}
-
-	public String getPackaging() {
-		return mPackaging;
-	}
-
-	public ValidAndInvalidClassifier getClassifier() {
-		return mClassifier;
-	}
-
 	WebResource getInstance() {
 		if (mInstance == null) {
 			init();
@@ -264,7 +240,7 @@ public class NexusLuceneSearchService implements IVersionReader {
 	public void setCredentials(String pUserName, String pUserPassword) {
 		this.setUserName(pUserName);
 		this.setUserPassword(pUserPassword);
-		
+
 	}
 
 }
