+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/central/MavenCentralChoiceListProvider/config.jelly
@@ -1,0 +1,50 @@
+<!--
+The MIT License
+
+Copyright (c) 2013 Stephan Watermeyer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-->
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+    <f:entry title="${%Credentials}" field="credentialsId">
+    	<c:select/>
+  	</f:entry>
+     <f:entry title="${%GroupId}" field="groupId">
+        <f:textbox />
+    </f:entry>
+     <f:entry title="${%ArtifactId}" field="artifactId">
+        <f:textbox />
+    </f:entry>
+     <f:entry title="${%Packaging}" field="packaging">
+        <f:textbox />
+    </f:entry>
+    <f:entry title="${%Classifier}" field="classifier">
+        <f:textbox />
+    </f:entry>
+     <f:entry title="${%Reverse Order}" field="reverseOrder">
+        <f:checkbox />
+    </f:entry>
+    
+    <f:validateButton
+        method="test"
+        title="${%List Files Now}"
+        progress="${%Checking...}"
+        with="credentialsId,groupId,artifactId,packaging,classifier,reverseOrder"
+    />
+</j:jelly>
