+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/central/MavenCentralChoiceListProvider/help-artifactId.html
