+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-artifactId.html
@@ -1,0 +1,3 @@
+<div>
+Mandatory: The artifactId to search for in the Nexus repository.
+</div>
