+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-classifier.html
@@ -1,0 +1,4 @@
+<div>
+The Maven Artifact classifier, i.E. "sources". By default leave empty.
+You can specifiy a list of valid or invalid classifier separated by Comma, i.E. "sources,!jar"
+</div>
