+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-credentialsId.html
@@ -1,0 +1,3 @@
+<div>
+If your repository service requires user authentication, you'll have to configure the username and password here.
+</div>
