+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-groupId.html
@@ -1,0 +1,3 @@
+<div>
+Optional: The GroupId to search the artifact, i.E. "com.wincor.components"
+</div>
