+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-packaging.html
@@ -1,0 +1,3 @@
+<div>Optional: Filters the matching packaging type from the
+	results, i.E. "jar" or "zip". Use a * to return all the values. If
+	parameter is not set, only the parent folder is returned.</div>
