+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-reverseOrder.html
@@ -1,0 +1,3 @@
+<div>
+If the list should be returned in reverse order.
+</div>
