+++ src/main/resources/org/jenkinsci/plugins/maven_artifact_choicelistprovider/nexus/NexusChoiceListProvider/help-url.html
@@ -1,0 +1,3 @@
+<div>
+Enter the URL of the Nexus server, i.E. http://nexus.mycompany.com/nexus/
+</div>
